moving extensions file
